UpdateFunctionalPreprocPackagesInPlace.sh now has a patch package only option

+++ MsmAllPackaging/UpdateFunctionalPreprocPackagesInPlace.sh
@@ -13,6 +13,7 @@ get_options()
     unset g_subject
     unset g_release_notes_template_file
     unset g_create_checksum
+	unset g_patch_only
 
     g_script_name=`basename ${0}`
 
@@ -49,6 +50,10 @@ get_options()
                 g_create_checksum="YES"
                 index=$(( index + 1 ))
                 ;;
+			--patch-only)
+				g_patch_only="TRUE"
+				index=$(( index + 1 ))
+				;;
             *)
                 echo "Unrecognized Option: ${argument}"
                 exit 1
@@ -101,6 +106,11 @@ get_options()
         g_create_checksum="NO"
     fi
     echo "create checksum: ${g_create_checksum}"
+
+	if [ -z "${g_patch_only}" ]; then
+		g_patch_only="FALSE"
+	fi
+	echo "patch_only: ${g_patch_only}"
 
     if [ ${error_count} -gt 0 ]; then
         echo "ERRORS DETECTED: EXITING"
@@ -272,40 +282,44 @@ main()
         
         popd
 
-        echo ""
-        echo "--------------------------------------------------"
-        echo " Update package"
-        echo "--------------------------------------------------"
-        echo ""
-        package_dir="${g_packages_root}/${g_subject}/preproc"
-        package_name="${g_subject}_3T_${scan}_preproc.zip"
-        package_path="${package_dir}/${package_name}"
+		if [ "${g_patch_only}" == "FALSE" ] ; then
 
-        # remove old checksum file
-        rm -rf ${package_path}.md5
+			echo ""
+			echo "--------------------------------------------------"
+			echo " Update package"
+			echo "--------------------------------------------------"
+			echo ""
+			package_dir="${g_packages_root}/${g_subject}/preproc"
+			package_name="${g_subject}_3T_${scan}_preproc.zip"
+			package_path="${package_dir}/${package_name}"
+		
+			# remove old checksum file
+			rm -rf ${package_path}.md5
+			
+			# update the zip file
+			pushd ${script_tmp_dir}
+			zip_cmd="zip --verbose --update --recurse-paths ${package_path} ${g_subject}"
+			echo "zip_cmd: ${zip_cmd}"
+			${zip_cmd}
 
-        # update the zip file
-        pushd ${script_tmp_dir}
-        zip_cmd="zip --verbose --update --recurse-paths ${package_path} ${g_subject}"
-        echo "zip_cmd: ${zip_cmd}"
-        ${zip_cmd}
+			# create the checksum file if requested
+			if [ "${g_create_checksum}" = "YES" ]; then
 
-        # create the checksum file if requested
-        if [ "${g_create_checksum}" = "YES" ]; then
+				echo ""
+				echo "--------------------------------------------------"
+				echo " Create MD5 Checksum"
+				echo "--------------------------------------------------"
+				echo ""
+				
+				pushd ${package_dir}
+				md5sum ${package_name} > ${package_name}.md5
+				chmod 777 ${package_name}.md5
+				popd
+			fi
+			
+			popd
 
-            echo ""
-            echo "--------------------------------------------------"
-            echo " Create MD5 Checksum"
-            echo "--------------------------------------------------"
-            echo ""
-            
-            pushd ${package_dir}
-            md5sum ${package_name} > ${package_name}.md5
-			chmod 777 ${package_name}.md5
-            popd
-        fi
-        
-        popd
+		fi
 
     done
